// src/lib/nobox-client/types.ts
export type StructureFieldType = typeof String | typeof Number | typeof Boolean | typeof Array | typeof Object;

export type CObject = Record<string, any>;

export interface StructureItem<T = any, K = StructureFieldType> {
  description?: string;
  comment?: string;
  required?: boolean;
  type: K;
  unique?: boolean;
  hashed?: boolean;
  defaultValue?: T;
}

export type Structure<T> = {
  [K in keyof T]: StructureItem<T[K]> | StructureFieldType;
};

export interface SpaceAuthOptions {
  active?: boolean;
  space: string;
  scope?: ('find' | 'insert' | 'delete' | 'update')[];
  token?: string;
}

export interface SpaceFunctionValues<T> {
  compulsoryParams: Partial<keyof T>[];
}

export interface SpaceFunctionOptions<T> {
  login?: SpaceFunctionValues<T>;
}

export interface SpaceWebhooks {
  onInsertUrl: string;
  onUpdateUrl: string;
}

export interface Space<T> {
  space: string;
  description: string;
  webhooks?: SpaceWebhooks;
  initialData?: T[];
  clear?: boolean;
  authOptions?: SpaceAuthOptions;
  structure: Structure<T>;
  functionOptions?: SpaceFunctionOptions<T>;
}